requirements.txt
Pillow==10.0.1
click==8.0.1
numpy
scikit-image